Fixes user page's user list
The 'LocationList' component was used in both the dashboard and the
users page, when it was changed for the dashboard it broke the user page.
This has been fixed by replicating the original LocationList component,
as it was before modifications, to a new component for use in the user page.